theme: fix invisible footnote element
Bootstrap's "label" CSS class is giving extra meaning to the
docutils "label" class, causing our footnote anchors to be made
invisible when there's more than one backref. Instead of making
label's meaning context-sensitive or trying to trick docutils into
applying an additional "label-info" class (which really gives us no
additional benefit and may not be possible anyway), I chose to just
rename Bootstrap's version of "label" to "bootstrap-label".

To the best of my knowledge we aren't using Bootstrap's label for
anything, but if we are, it's a trivial update.

Signed-off-by: Sam Kleinman <samk@10gen.com>