DOCSP-4263 Rollup standalone feedback widget
This creates a new css and js target for the feedback widget alone,
so that it can be consumed by SDK generated docs themes.

The entry point directory main/ is created so that entrypoints
are not globbed into each others' bundles. Specifically, the new
feedbackController.js would be globbed into the controller.js bundle.